+++ contracts/mocks/token/ERC20/ChildERC20Mock.sol
@@ -2,20 +2,28 @@
 
 pragma solidity >=0.7.6 <0.8.0;
 
-import {ManagedIdentity, Ownable, Recoverable, RecoverableERC20} from "@animoca/ethereum-contracts-core-1.0.1/contracts/utils/Recoverable.sol";
+import {IWrappedERC20, ERC20Wrapper} from "@animoca/ethereum-contracts-core-1.1.0/contracts/utils/ERC20Wrapper.sol";
+import {ManagedIdentity, Ownable, Recoverable} from "@animoca/ethereum-contracts-core-1.1.0/contracts/utils/Recoverable.sol";
 import {IForwarderRegistry, UsingUniversalForwarding} from "ethereum-universal-forwarder/src/solc_0.7/ERC2771/UsingUniversalForwarding.sol";
 import {ChildERC20} from "../../../token/ERC20/ChildERC20.sol";
 import {IERC20Mintable} from "../../../token/ERC20/IERC20Mintable.sol";
 
 contract ChildERC20Mock is Recoverable, UsingUniversalForwarding, ChildERC20, IERC20Mintable {
+    using ERC20Wrapper for IWrappedERC20;
+
     uint256 internal _inEscrow;
 
     constructor(
         address[] memory recipients,
         uint256[] memory values,
+        address childChainManager,
         IForwarderRegistry forwarderRegistry,
         address universalForwarder
-    ) ChildERC20("Child ERC20 Mock", "CE20", 18, "uri") Ownable(msg.sender) UsingUniversalForwarding(forwarderRegistry, universalForwarder) {
+    )
+        ChildERC20("Child ERC20 Mock", "CE20", 18, "uri", childChainManager)
+        Ownable(msg.sender)
+        UsingUniversalForwarding(forwarderRegistry, universalForwarder)
+    {
         _batchMint(recipients, values);
     }
 
@@ -33,10 +41,8 @@ contract ChildERC20Mock is Recoverable, UsingUniversalForwarding, ChildERC20, IE
     }
 
     function deposit(address user, bytes calldata depositData) public virtual override {
-        _requireDepositorRole(_msgSender());
-        uint256 amount = abi.decode(depositData, (uint256));
-        _inEscrow -= amount;
-        _transfer(address(this), user, amount);
+        _inEscrow -= abi.decode(depositData, (uint256));
+        super.deposit(user,depositData);
     }
 
     function withdraw(uint256 amount) public virtual override {
@@ -69,8 +75,13 @@ contract ChildERC20Mock is Recoverable, UsingUniversalForwarding, ChildERC20, IE
                 uint256 recoverable = _balances[address(this)] - _inEscrow;
                 require(amount <= recoverable, "Recov: insufficient balance");
             }
-            require(RecoverableERC20(token).transfer(accounts[i], amount), "Recov: transfer failed");
+            IWrappedERC20(token).wrappedTransfer(accounts[i], amount);
         }
+    }
+
+    function setTokenURI(string calldata tokenURI_) external {
+        _requireOwnership(_msgSender());
+        _tokenURI = tokenURI_;
     }
 
     /**