+++ contracts/mocks/token/ERC20/ERC20BurnableMock.sol
@@ -2,7 +2,7 @@
 
 pragma solidity >=0.7.6 <0.8.0;
 
-import {ManagedIdentity, Ownable, Recoverable} from "@animoca/ethereum-contracts-core-1.0.1/contracts/utils/Recoverable.sol";
+import {ManagedIdentity, Ownable, Recoverable} from "@animoca/ethereum-contracts-core-1.1.0/contracts/utils/Recoverable.sol";
 import {IForwarderRegistry, UsingUniversalForwarding} from "ethereum-universal-forwarder/src/solc_0.7/ERC2771/UsingUniversalForwarding.sol";
 import {ERC20Burnable} from "../../../token/ERC20/ERC20Burnable.sol";
 import {IERC20Mintable} from "../../../token/ERC20/IERC20Mintable.sol";
@@ -28,6 +28,11 @@ contract ERC20BurnableMock is Recoverable, UsingUniversalForwarding, ERC20Burnab
 
     function _msgData() internal view virtual override(ManagedIdentity, UsingUniversalForwarding) returns (bytes memory ret) {
         return UsingUniversalForwarding._msgData();
+    }
+
+    function setTokenURI(string calldata tokenURI_) external {
+        _requireOwnership(_msgSender());
+        _tokenURI = tokenURI_;
     }
 
     /**