+++ hardhat-config/imports.config.js
@@ -1,4 +1,7 @@
 module.exports = {
-  imports: ['node_modules/@animoca/ethereum-contracts-core-1.0.1/artifacts'],
-  imports: ['node_modules/ethereum-universal-forwarder/export/artifacts'],
+  imports: [
+    'node_modules/@animoca/ethereum-contracts-core-1.1.0/artifacts',
+    'node_modules/ethereum-universal-forwarder/export/artifacts',
+    'artifacts_pos-portal',
+  ],
 };