+++ test/contracts/token/ERC20/ChildERC20BurnableMock.test.js
@@ -25,11 +25,17 @@ const implementation = {
     TransferRefused: 'ERC20: transfer refused',
     MintToZero: 'ERC20: zero address',
     BatchMintValuesOverflow: 'ERC20: values overflow',
+    BurnFromZero: 'ERC20: insufficient balance',
+    BurnExceedsBalance: 'ERC20: insufficient balance',
+    BurnExceedsAllowance: 'ERC20: insufficient allowance',
+    BatchBurnValuesOverflow: 'ERC20: insufficient balance',
     SupplyOverflow: 'ERC20: supply overflow',
     PermitFromZero: 'ERC20: zero address owner',
     PermitExpired: 'ERC20: expired permit',
     PermitInvalid: 'ERC20: invalid permit',
     NonMinter: 'Ownable: not the owner',
+    NonDepositor: 'ChildERC20: only depositor',
+    DirectReceiverCall: 'ChildERC20: wrong sender',
   },
   features: {
     ERC165: true,
@@ -44,8 +50,20 @@ const implementation = {
     ERC20BatchTransfer: true,
     ERC20Safe: true,
     ERC20Permit: true,
+    ChildToken: true,
   },
   methods: {
+    // ERC20Burnable
+    'burn(uint256)': async (contract, value, overrides) => {
+      return contract.burn(value, overrides);
+    },
+    'burnFrom(address,uint256)': async (contract, from, value, overrides) => {
+      return contract.burnFrom(from, value, overrides);
+    },
+    'batchBurnFrom(address[],uint256[])': async (contract, owners, values, overrides) => {
+      return contract.batchBurnFrom(owners, values, overrides);
+    },
+
     // ERC20Mintable
     'mint(address,uint256)': async (contract, account, value, overrides) => {
       return contract.mint(account, value, overrides);
@@ -55,13 +73,16 @@ const implementation = {
     },
   },
   deploy: async function (initialHolders, initialBalances, deployer) {
-    const forwarder = await artifacts.require('UniversalForwarder').new({from: deployer});
     const registry = await artifacts.require('ForwarderRegistry').new({from: deployer});
-    return artifacts.require('ChildERC20BurnableMock').new(initialHolders, initialBalances, forwarder.address, registry.address, {from: deployer});
+    const forwarder = await artifacts.require('UniversalForwarder').new({from: deployer});
+    const childChainManager = deployer;
+    return artifacts
+      .require('ChildERC20BurnableMock')
+      .new(initialHolders, initialBalances, childChainManager, registry.address, forwarder.address, {from: deployer});
   },
 };
 
-describe('ERC20Mock', function () {
+describe('ChildERC20BurnableMock', function () {
   this.timeout(0);
 
   const [deployer, other] = accounts;