+++ test/contracts/token/ERC20/ChildERC20Mock.test.js
@@ -30,6 +30,8 @@ const implementation = {
     PermitExpired: 'ERC20: expired permit',
     PermitInvalid: 'ERC20: invalid permit',
     NonMinter: 'Ownable: not the owner',
+    NonDepositor: 'ChildERC20: only depositor',
+    DirectReceiverCall: 'ChildERC20: wrong sender',
   },
   features: {
     ERC165: true,
@@ -44,6 +46,7 @@ const implementation = {
     ERC20BatchTransfer: true,
     ERC20Safe: true,
     ERC20Permit: true,
+    ChildToken: true,
   },
   methods: {
     // ERC20Mintable
@@ -55,13 +58,16 @@ const implementation = {
     },
   },
   deploy: async function (initialHolders, initialBalances, deployer) {
-    const forwarder = await artifacts.require('UniversalForwarder').new({from: deployer});
     const registry = await artifacts.require('ForwarderRegistry').new({from: deployer});
-    return artifacts.require('ChildERC20Mock').new(initialHolders, initialBalances, forwarder.address, registry.address, {from: deployer});
+    const forwarder = await artifacts.require('UniversalForwarder').new({from: deployer});
+    const childChainManager = deployer;
+    return artifacts
+      .require('ChildERC20Mock')
+      .new(initialHolders, initialBalances, childChainManager, registry.address, forwarder.address, {from: deployer});
   },
 };
 
-describe('ERC20Mock', function () {
+describe('ChildERC20Mock', function () {
   this.timeout(0);
 
   const [deployer, other] = accounts;